rename exceptions.py module to errors.py
