+++ tests/test_middleware.py
@@ -1,7 +1,7 @@
 from unittest import mock
 
 import middlewares
-import exceptions
+import errors
 
 
 async def handler_func(request, *args, **kwargs):
@@ -21,17 +21,17 @@ async def test_error_middleware_500():
 
 async def test_error_middleware_400():
     with mock.patch('middlewares.handle_http_error', handle_http_error):
-        res = await middlewares.error_middleware(request=exceptions.BadRequest, handler=handler_func)
+        res = await middlewares.error_middleware(request=errors.BadRequest, handler=handler_func)
     assert res == 400
 
 
 async def test_error_middleware_404():
     with mock.patch('middlewares.handle_http_error', handle_http_error):
-        res = await middlewares.error_middleware(request=exceptions.RecordNotFound, handler=handler_func)
+        res = await middlewares.error_middleware(request=errors.RecordNotFound, handler=handler_func)
     assert res == 404
 
 
 async def test_error_middleware_403():
     with mock.patch('middlewares.handle_http_error', handle_http_error):
-        res = await middlewares.error_middleware(request=exceptions.UserIsNotActivated, handler=handler_func)
+        res = await middlewares.error_middleware(request=errors.UserIsNotActivated, handler=handler_func)
     assert res == 403
